% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraNumericInput.R
\name{chakraNumberInput}
\alias{chakraNumberInput}
\title{Chakra number input}
\usage{
chakraNumberInput(
  inputId,
  label = NULL,
  value,
  min,
  max,
  step = NULL,
  size = "md",
  numberInputOptions = list()
)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value}

\item{label}{the label for the widget; this can be some HTML code}

\item{value}{initial value}

\item{min}{minimum allowed value}

\item{max}{maximum allowed value}

\item{step}{stepping interval to use when adjusting the value}

\item{size}{size of the widget, can be \code{"sm"} (small),
\code{"md"} (medium) or \code{"lg"} (large)}

\item{numberInputOptions}{list of options for the number input;
see \code{\link{numberInputOptions}}}
}
\description{
This creates a number input in the Shiny UI.
}
