% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{chakraSliderInput}
\alias{chakraSliderInput}
\title{<Add Title>}
\usage{
chakraSliderInput(
  inputId,
  value,
  min,
  max,
  step = NULL,
  width = "100\%",
  size = "md",
  numberInputOptions = list(),
  trackColor = NULL,
  thumbOptions = list(),
  gap = "2rem"
)
}
\description{
<Add Description>
}
