% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{numberInputOptions}
\alias{numberInputOptions}
\title{Options for the number input of \code{chakraSliderInput}}
\usage{
numberInputOptions(
  width = NULL,
  fontSize = NULL,
  fontColor = NULL,
  borderColor = NULL,
  focusBorderColor = NULL,
  borderWidth = NULL,
  stepperColor = NULL
)
}
\description{
Create a list of options to be passed to
\code{numberInputOptions} in \code{chakraSliderInput}.
}
