% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraNumericInput.R
\name{updateChakraNumberInput}
\alias{updateChakraNumberInput}
\title{Update a chakra number input}
\usage{
updateChakraNumberInput(session, inputId, value)
}
\arguments{
\item{session}{the Shiny session object}

\item{inputId}{the id of the chakra number input to update}

\item{value}{the new value of the chakra number input}
}
\description{
Update the value of a chakra number input.
}
