% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chakraSlider.R
\name{thumbOptions}
\alias{thumbOptions}
\title{Options for the thumb of \code{chakraSliderInput}}
\usage{
thumbOptions(
  width = NULL,
  height = NULL,
  color = NULL,
  borderColor = NULL,
  borderWidth = NULL,
  icon = NULL,
  iconColor = NULL,
  iconSize = NULL
)
}
\description{
Create a list of options to be passed to \code{thumbOptions}
in \code{chakraSliderInput}.
}
